refactor(porwr): group components by surface, one convention
`src/components/` had 21 files flat at the top level beside seven
directories following two different conventions: PascalCase folders
holding a single `Index.vue`, and kebab-case folders grouped by surface
or feature. Now there is one: `sidebar/`, `dashboard/`, `extension/`,
`shared/`, and nothing at the top level.

Placement was decided by walking the import graph from every routed page
and layout rather than by name. Three components are reachable from more
than one surface and go to `shared/`: DiogelLogo (all three),
AccountDropdown (dashboard and extension), and ProfileView (sidebar and
extension). ProfileView is the one that motivated the directory — it
renders identity in two places that have to agree, so duplicating it was
never an option.

The four `X/Index.vue` folders are now plain files named for what they
are, which is what stack traces and editor tabs wanted all along. The
one colocated spec moved to `tests/unit/components/`, leaving no tests
under `src/`.

Every import in the moved files is an alias now. Relative imports were
the whole hazard of this change: `../types` meant `src/types` from
`src/components/X.vue` and would mean `src/components/types` one level
down, and it fails as a missing module rather than as anything
describing the cause.

Refs #180
