+++ src/components/dashboard/DashboardFooter.vue
