+++ src/components/dashboard/DashboardNavigation.vue
@@ -3,7 +3,7 @@ import { useRouter } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 import { useVault } from 'src/composables/useVault';
 import { useNavigation } from 'src/composables/useNavigation';
-import DiogelLogo from 'components/DiogelLogo/Index.vue';
+import DiogelLogo from 'components/shared/DiogelLogo.vue';
 import type { NavigationItem, UtilityLinkItem } from 'src/types/navigation';
 
 const { handleLock } = useVault();
