+++ src/components/dashboard/ImagePreview.vue
