+++ src/components/dashboard/ProfileImage.vue
@@ -2,12 +2,12 @@
 import { onMounted, ref, watch, computed } from 'vue';
 import { useQuasar } from 'quasar';
 import { useI18n } from 'vue-i18n';
-import type { NostrProfile, StoredKey } from '../types';
+import type { NostrProfile, StoredKey } from 'src/types';
 import { finalizeEvent, getPublicKey, SimplePool } from 'nostr-tools';
 import { hexToBytes } from '@noble/hashes/utils';
-import useSettingsStore from '../stores/settings-store';
-import BannerEditor from './BannerEditor.vue';
-import AvatarEditor from './AvatarEditor.vue';
+import useSettingsStore from 'src/stores/settings-store';
+import BannerEditor from 'components/dashboard/BannerEditor.vue';
+import AvatarEditor from 'components/dashboard/AvatarEditor.vue';
 
 defineOptions({ name: 'ProfileImage' });
