+++ src/components/dashboard/ViewStoredKey.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import type { StoredKey } from '../../types';
+import type { StoredKey } from 'src/types/';
 import { useQuasar } from 'quasar';
 import { computed, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
