+++ src/components/extension/ExtensionNavigation.vue
