+++ src/components/extension/ExtensionWindowHeader.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
-import AccountDropdown from 'components/AccountDropdown/Index.vue';
-import ExtensionNavigation from 'components/ExtensionNavigation.vue';
+import AccountDropdown from 'components/shared/AccountDropdown.vue';
+import ExtensionNavigation from 'components/extension/ExtensionNavigation.vue';
 import useVaultStore from 'src/stores/vault-store';
 
 const vaultStore = useVaultStore();
