+++ src/pages/dashboard/DashboardPage.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
-import ActiveKeysCard from 'src/components/dashboard/ActiveKeysCard.vue';
-import ConnectedRelaysCard from 'src/components/dashboard/ConnectedRelaysCard.vue';
-import QuickSignCard from 'src/components/dashboard/QuickSignCard.vue';
-import RecentActivityCard from 'src/components/dashboard/RecentActivityCard.vue';
-import ApprovedClientsCard from 'src/components/dashboard/ApprovedClientsCard.vue';
+import ActiveKeysCard from 'components/dashboard/ActiveKeysCard.vue';
+import ConnectedRelaysCard from 'components/dashboard/ConnectedRelaysCard.vue';
+import QuickSignCard from 'components/dashboard/QuickSignCard.vue';
+import RecentActivityCard from 'components/dashboard/RecentActivityCard.vue';
+import ApprovedClientsCard from 'components/dashboard/ApprovedClientsCard.vue';
 import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
 import { getDashboardSummary } from 'src/services/dashboard-service';
