+++ src/pages/dashboard/KeyManagementPage.vue
@@ -2,8 +2,8 @@
 import { computed, onMounted } from 'vue';
 import { useI18n } from 'vue-i18n';
 import useAccountStore from 'src/stores/account-store';
-import KeyManagementTable from 'components/key-management/KeyManagementTable.vue';
-import SecurityWarning from 'components/SecurityWarning.vue';
+import KeyManagementTable from 'components/dashboard/key-management/KeyManagementTable.vue';
+import SecurityWarning from 'components/dashboard/SecurityWarning.vue';
 
 const { t } = useI18n();
 const accountStore = useAccountStore();
