+++ src/pages/dashboard/ProfilePage.vue
@@ -3,9 +3,9 @@ import { computed, onMounted, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRoute, useRouter } from 'vue-router';
 import useAccountStore from 'src/stores/account-store';
-import ProfileImage from 'components/ProfileImage.vue';
-import ProfileEditor from 'components/ProfileEditor.vue';
-import ProfilePreview from 'components/ProfilePreview.vue';
+import ProfileImage from 'components/dashboard/ProfileImage.vue';
+import ProfileEditor from 'components/dashboard/ProfileEditor.vue';
+import ProfilePreview from 'components/dashboard/ProfilePreview.vue';
 
 
 const { t } = useI18n();
