+++ src/pages/extension/IndexPage.vue
@@ -22,7 +22,7 @@
 import { computed, onMounted } from 'vue';
 import { useRouter } from 'vue-router';
 import useAccountStore from 'src/stores/account-store';
-import ProfileView from 'components/ProfileView.vue';
+import ProfileView from 'components/shared/ProfileView.vue';
 
 const accountStore = useAccountStore();
 const router = useRouter();
