+++ tests/unit/components/GenerateKeyForm.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { mount } from '@vue/test-utils';
 import { nextTick } from 'vue';
 
-import GenerateKeyForm from 'src/components/key-management/GenerateKeyForm.vue';
+import GenerateKeyForm from 'components/dashboard/key-management/GenerateKeyForm.vue';
 
 const { saveKeyMock, pushMock, notifyMock, generateKeyMock } = vi.hoisted(() => ({
   saveKeyMock: vi.fn(),
