+++ tests/unit/components/ImportKeyForm.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { mount } from '@vue/test-utils';
 import { nextTick } from 'vue';
 
-import ImportKeyForm from 'src/components/key-management/ImportKeyForm.vue';
+import ImportKeyForm from 'components/dashboard/key-management/ImportKeyForm.vue';
 
 const { saveKeyMock, pushMock, notifyMock } = vi.hoisted(() => ({
   saveKeyMock: vi.fn(),
